requirements.txt
pandas~=2.1.3
PyQt6~=6.6.1
PyQt6_sip==13.6.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
mysql~=0.0.3
mysql-connector-python~=8.3.0
mysql-connector